# docs/openapi-frontend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MeController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/guilds": {
      "get": {
        "operationId": "MeController_getMyGuilds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/team": {
      "get": {
        "operationId": "MeController_getTeam",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/team/invite": {
      "post": {
        "operationId": "MeController_inviteTeamMember",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/team/{memberId}": {
      "patch": {
        "operationId": "MeController_updateTeamMember",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTeamMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MeController_removeTeamMember",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/subscription": {
      "get": {
        "operationId": "MeController_getSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/subscription/upgrade": {
      "post": {
        "operationId": "MeController_upgradeSubscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/subscription/cancel": {
      "post": {
        "operationId": "MeController_cancelSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/usage": {
      "get": {
        "operationId": "MeController_getUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/invoices": {
      "get": {
        "operationId": "MeController_getInvoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/invoices/{invoiceId}/download": {
      "get": {
        "operationId": "MeController_downloadInvoice",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds": {
      "get": {
        "operationId": "GuildsController_getGuilds",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Guilds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/guilds/{guildId}/widgets": {
      "get": {
        "operationId": "WidgetsController_getWidgets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Widgets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "WidgetsController_createWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWidgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Widgets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/guilds/{guildId}/widgets/{widgetId}": {
      "patch": {
        "operationId": "WidgetsController_updateWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchWidgetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Widgets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "WidgetsController_deleteWidget",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Widgets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/users/me/data": {
      "delete": {
        "operationId": "UserDataController_deleteAllData",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "User Data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/me/data/export": {
      "get": {
        "operationId": "UserDataController_exportData",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Frontend API",
    "description": "Public API. Authentication: Bearer JWT in Authorization header. All requests/responses use application/json. Envelope pattern: success in data, errors in error.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InviteTeamDto": {
        "type": "object",
        "properties": {}
      },
      "PatchTeamMemberDto": {
        "type": "object",
        "properties": {}
      },
      "UpgradeSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
      "PatchMeDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateWidgetDto": {
        "type": "object",
        "properties": {}
      },
      "PatchWidgetDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}